+++ src/main/ruby/truffleruby/core/thread.rb
@@ -411,23 +411,13 @@ class Thread
     Truffle::System.synchronized(self) { @thread_local_variables.keys }
   end
 
+  def backtrace(omit = 0, length = undefined)
+    omit, length = Truffle::KernelOperations.normalize_backtrace_args(omit, length)
+    Primitive.thread_backtrace(self, omit, length)
+  end
+
   def backtrace_locations(omit = 0, length = undefined)
-    if Integer === length && length < 0
-      raise ArgumentError, "negative size (#{length})"
-    end
-    if Range === omit
-      range = omit
-      omit = Truffle::Type.coerce_to_int(range.begin)
-      unless range.end.nil?
-        end_index = Truffle::Type.coerce_to_int(range.end)
-        if end_index < 0
-          length = end_index
-        else
-          end_index += (range.exclude_end? ? 0 : 1)
-          length = omit > end_index ? 0 : end_index - omit
-        end
-      end
-    end
+    omit, length = Truffle::KernelOperations.normalize_backtrace_args(omit, length)
     Primitive.thread_backtrace_locations(self, omit, length)
   end
 end